git-svn-id: svn://svn.r-forge.r-project.org/svnroot/casper@84 edb9625f-4e0d-4859-8d74-9fd3b1da38cb
Former-commit-id: 2a1e9b28fadbe5c2bfccfa9654f3be47a3271cd3

+++ pkg/casper/src/rcasper.cpp
@@ -299,7 +299,7 @@ char* getmodelcode(vector<Variant*>* allvariants, Model* model)
 	char* str = new char[n + 1];
 	str[n] = '\0';
 
-	for (int i = 0; i < allvariants->size(); i++)
+	for (int i = 0; i < (int)allvariants->size(); i++)
 	{
 		if (model->contains(allvariants->at(i)))
 		{
